# Make TF pt-tf equivalence test more aggressive