--- a/tests/test_modeling_tf_common.py
+++ b/tests/test_modeling_tf_common.py
@@ -467,15 +467,10 @@ class TFModelTesterMixin:
                 if "TFConvNext" not in model_class.__name__:
                     output_kwargs["output_attentions"] = True
 
-                tf_inputs_dict.update(output_kwargs)
-                pt_inputs_dict.update(output_kwargs)
-                tf_inputs_dict_maybe_with_labels.update(output_kwargs)
-                pt_inputs_dict_maybe_with_labels.update(output_kwargs)
-
                 # Original test: check without `labels`
                 with torch.no_grad():
-                    pto = pt_model(**pt_inputs_dict)
-                tfo = tf_model(tf_inputs_dict)
+                    pto = pt_model(**pt_inputs_dict, **output_kwargs)
+                tfo = tf_model(tf_inputs_dict, **output_kwargs)
 
                 tf_keys = [k for k, v in tfo.items() if v is not None]
                 pt_keys = [k for k, v in pto.items() if v is not None]
@@ -490,8 +485,8 @@ class TFModelTesterMixin:
                 if has_labels:
 
                     with torch.no_grad():
-                        pto = pt_model(**pt_inputs_dict_maybe_with_labels)
-                    tfo = tf_model(tf_inputs_dict_maybe_with_labels)
+                        pto = pt_model(**pt_inputs_dict_maybe_with_labels, **output_kwargs)
+                    tfo = tf_model(tf_inputs_dict_maybe_with_labels, **output_kwargs)
 
                     # Some models' output class don't have `loss` attribute despite `labels` is used.
                     # TODO: identify which models