… check_speed.py