# What does this PR do?

Make TF pt-tf equivalence test more aggressive.

After a series of fixes done so far, I think it is a good time to include this more aggressive testing in `master` branch.
(Otherwise, the new models added might have undetected issues. For example, the recent `TFConvNextModel` would have `hidden_states` not transposed to match Pytorch version issue - I tested it on my local branch and informed the author to fix it).

There are still 3 categories of PT/TF inconsistency to address, but they are less urgent in my opinion. See below.
Currently, the test makes a few exception to not test these 3 cases (in order to get green test) - I add `TODO` comments in the code.

TF: @Rocketknight1 @gante 
Test: @LysandreJik  @sgugger 

## TODO in separate PRs

- failing due to the difference of large negative values (used for attn mask) between PT/TF:
  - https://circleci.com/api/v1.1/project/github/huggingface/transformers/379126/output/112/0?file=true&allocation-id=621b51ac72134246517cac0c-0-build%2F6E4C1534
  - albert, convbert, speech_to_text, t5, tapas 
- failing due to cache format difference between PT/TF:
  - https://app.circleci.com/pipelines/github/huggingface/transformers/35203/workflows/d0c0fbb4-661b-4a61-96e7-2fd4a3249f44/jobs/379098/parallel-runs/0/steps/0-112
  - gpt2, t5, speech_to_text
- failing due to outputting loss or not between PT/TF:
  - https://app.circleci.com/pipelines/github/huggingface/transformers/35206/workflows/d2533f38-6877-4db5-9eb6-da92e6421e11/jobs/379144/parallel-runs/0/steps/0-112
  - flaubert, funnel, transfo_xl, xlm