@ydshieh 2022-02-28
> Thank you for your time spent making this better! Two questions:
> 
>     * Does it run on GPU?

    For now, I haven't tested it with GPU. I can run it on the office's  GPU machine this week (a good chance to learn how to connect to those machine!)

> 
>     * How long does it take to run? Our test suite is already taking a significant time to run, so aiming for common tests that run as fast as possible is important.

   Let me measure the timing of this test with the current master and with this PR. Will report it.

> 
> 
> There's quite a bit of model-specific logic which I'm not particularly enthusiastic about (pre-training models + convnext), but I understand why it's rigorous to do it like that here.