@ydshieh 2022-03-01
For the test timing:

| Model    |      Before (s)    | This PR (s) |
|----------|:-------------:|------:|
| BERT | 7.68  | 6.12 |
| GPT2 |    2.64   | 1.78  |
| BART | 4.69 |    3.52 |
| T5 |  2.92 |  1.87 |
| Total |  4m37 |  5m22 |

It's somehow strange that the new test took less time to run for each above model - although it had a longer total run time.
Do you want me to digger deeper for this, @LysandreJik ?
(one )

[Report before this PR](https://app.circleci.com/pipelines/github/huggingface/transformers/35311/workflows/cf12b350-c7ca-4382-82ea-63c418488620/jobs/380708)

[Report after this PR](https://app.circleci.com/pipelines/github/huggingface/transformers/35231/workflows/001c1bc0-f0bf-4bc6-9823-afb54f4e927d/jobs/379503)