@ydshieh 2022-03-01
For this testing on GPU (running on a GCP VM with 1 GPU) 
```
== 143 passed, 5 skipped, 16494 deselected, 16 warnings in 352.17s (0:05:52) ===
```

(after an update to make pytorch models/inputs on the correct device)

This looks good, i.e `1e-5` still works on GPU, at least for the current models

However, I would like to make sure again the PyTorch models and inputs DOES run on GPU by stepping into the execution (I don't like to be fooled by potential mistakes). Will confirm once it is done.