@ydshieh 2022-03-01
@LysandreJik 

I run the previous test and the new test on GCP CPU VM. Here is what I obtained so far.
(There are full output files at the end)

## Observation

- testing is faster in this PR (I can dig deeper, say running `TFBertModelTest::test_pt_tf_model_equivalence` manually)
- the timing (per model testing) recorded with `-n 8` seems very strange, but the final timing looks reasonable. (I don't know why), but it is not very performing (with 8 workers is only 2x faster than only 1 worker)

## Command I use

```
RUN_PT_TF_CROSS_TESTS=1 python -m pytest -n 1 -v tests -k pt_tf_model_equivalence --durations=0 | tee output.txt
```

## Full outputs with some summary

- Before: previous testing method
- After = new testing method done in this PR

----

- 1 worker (-n 1)

  - before this PR: (333.26s)
[test_pt_tf_model_equivalence_cpu_n1_output_commit_84eaa6acf582206dba33135727dc3bfff05a7e9c_before.txt](https://github.com/huggingface/transformers/files/8164548/test_pt_tf_model_equivalence_cpu_n1_output_commit_84eaa6acf582206dba33135727dc3bfff05a7e9c_before.txt)

    - TFBertModelTest::test_pt_tf_model_equivalence: 7.30s
    - TFGPT2ModelTest::test_pt_tf_model_equivalence: 2.72s

  - after this PR: (241.08s)
[test_pt_tf_model_equivalence_cpu_n1_output_commit_cce5b6a6511880b5c6d7dc1b74d327485c595ffc_after.txt](https://github.com/huggingface/transformers/files/8164551/test_pt_tf_model_equivalence_cpu_n1_output_commit_cce5b6a6511880b5c6d7dc1b74d327485c595ffc_after.txt)

    - TFBertModelTest::test_pt_tf_model_equivalence: 3.44s
    - TFGPT2ModelTest::test_pt_tf_model_equivalence: 1.10s

- 8 workers (-n 8)

  - before this PR: (167.67s)
[test_pt_tf_model_equivalence_cpu_n8_output_commit_84eaa6acf582206dba33135727dc3bfff05a7e9c_before.txt](https://github.com/huggingface/transformers/files/8164558/test_pt_tf_model_equivalence_cpu_n8_output_commit_84eaa6acf582206dba33135727dc3bfff05a7e9c_before.txt)


    - TFBertModelTest::test_pt_tf_model_equivalence: 29.80s (very strange values!)
    - TFGPT2ModelTest::test_pt_tf_model_equivalence: 11.22s

  - after this PR: (153.71s)
[test_pt_tf_model_equivalence_cpu_n8_output_commit_cce5b6a6511880b5c6d7dc1b74d327485c595ffc_after.txt](https://github.com/huggingface/transformers/files/8164561/test_pt_tf_model_equivalence_cpu_n8_output_commit_cce5b6a6511880b5c6d7dc1b74d327485c595ffc_after.txt)


    - TFBertModelTest::test_pt_tf_model_equivalence: 26.40ss
    - TFGPT2ModelTest::test_pt_tf_model_equivalence: 6.08s