@ydshieh 2022-03-02
> It's weird that it took more time when you expected it to take less, no? Can you try running the test suite with `--durations=0` to see all the tests and the time it took for them to run?

Sure! I actually copied the 2 results from CircleCI report. But I will run it on GCP on CPU to measure the timing.