@ydshieh 2022-03-02
After a deeper investigation, I realized that I made a mistake (accidentally removed a test case in the original version).
This is why the observed running time was shorter than before.

In the original test, there were 2 parts:

- https://github.com/huggingface/transformers/blob/2eb7bb15e771f13192968cd4657c78f76b0799fe/tests/test_modeling_tf_common.py#L362
- https://github.com/huggingface/transformers/blob/2eb7bb15e771f13192968cd4657c78f76b0799fe/tests/test_modeling_tf_common.py#L401

The last commit fixed it, and now the correct running time is:

| original test | new test |
| -------------:| -------------:|
| 54.85s  | 66.77s  |

@LysandreJik Let me know if this increasement is acceptable, thanks.