@ydshieh 2022-03-05
After a thorough verification and a few fixes, this PR is ready (again) from my side.

I would love sgugger (and Lysandre when he is back) to check it again, and gante & Rocketknight1 if they want (no particular TF-related change compared to the last commit).

The following summary might save some review (again) time

- **Main fixes** (since the last review):

  - In the last commit, the following line forgot to use `to_tuple()`, and the method `check_output` (previous version) only deal with `tuple` or `tensor`. So half test cases passed just because no check was performed at all. (my bad ...).
https://github.com/huggingface/transformers/blob/6f2025054dc49a42dc3de82cbea22f2a5913f122/tests/test_modeling_tf_common.py#L447
This is fixed now, and also `check_output` now raises error if a result is not tested (if its type is not in `[tuple, list, tensor]`).

  - In the last commit, I accidentally deleted the following test case (`check the results after save/load checkpoints`, which exists on the current `master`).
https://github.com/huggingface/transformers/blob/6f2025054dc49a42dc3de82cbea22f2a5913f122/tests/test_modeling_tf_common.py#L576
This is added back now.

  - Make the test also run on GPU (put models/inputs on the correct device)

- **Style changes**

  - Since I have to add back the accidentally deleted test cases , I added `check_pt_tf_models` to avoid duplicated large code block.
  - `tfo` -> `tf_outputs` and `pto` -> `pt_outputs`

I run this new version of test, and with the small tolerance `1e-5`, it pass both on CPU and GPU :tada:!

In order to be super sure, I also ran it on CPU/GPU for 100 times (very aggressive :fire: :fire: )! All models passed for this test, except `TFSpeech2TextModel`. See ...

Regarding the running time, I will measure it and post the results in the next comment.